only filter runTests command. Make it a flag and optional. Run only changed tests for travis
